fix: setting for design
